WebUI performance optimization and settings schema audit
Fix config save failure after service deploy by chowning dawn.toml
to the service user. Show local LLM settings (endpoint, model, vision)
without requiring Advanced toggle for initial setup.

Eliminate all build warnings: add (long) casts for curl timeout args,
replace deprecated CURLOPT_PROTOCOLS with version-guarded compat macro
(DAWN_CURL_SET_PROTOCOLS) that falls back to bitmask parsing on
curl < 7.85.0 for Jetson compatibility, fix realloc size-proof warning
in webui_documents.c.

Add memory extraction fallback: when the configured extraction model
fails (e.g. HTTP 404), retry with the session's active model and
notify the user via a transient WebUI banner. Shared fallback builder
uses session_get_llm_config() for proper mutex acquisition across all
5 call sites.

Fix thread-safety race on g_config.network.llm_timeout_ms: extraction
and summarization threads mutated the global without synchronization.
Replaced with thread-local timeout override (llm_get_effective_timeout_ms)
read by llm_openai.c and llm_claude.c, eliminating cross-thread
corruption of timeout values.

+++ src/tools/search_summarizer.c
@@ -291,8 +291,8 @@ static int summarize_with_local_llm(const char *prompt, char **out_summary) {
    curl_easy_setopt(curl, CURLOPT_POSTFIELDS, payload);
    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, curl_buffer_write_callback);
    curl_easy_setopt(curl, CURLOPT_WRITEDATA, &buffer);
-   curl_easy_setopt(curl, CURLOPT_TIMEOUT, SUMMARIZER_LOCAL_TIMEOUT_SEC);
-   curl_easy_setopt(curl, CURLOPT_CONNECTTIMEOUT, 5);
+   curl_easy_setopt(curl, CURLOPT_TIMEOUT, (long)SUMMARIZER_LOCAL_TIMEOUT_SEC);
+   curl_easy_setopt(curl, CURLOPT_CONNECTTIMEOUT, 5L);
 
    LOG_INFO("search_summarizer: Sending %zu byte prompt to local LLM", strlen(prompt));
